Move pulp_file tests
[noissue]
