--- pulp_file/tests/functional/api/from_pulpcore/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for core functionality that require plugin involvement to exercise."""
