--- pulp_file/tests/unit/test_models.py
@@ -1,9 +1,0 @@
-from django.test import TestCase
-
-
-class TestNothing(TestCase):
-    """Test Nothing (placeholder)."""
-
-    def test_nothing_at_all(self):
-        """Test that the tests are running and that's it."""
-        self.assertTrue(True)
