+++ pulpcore/tests/functional/api/pulp_file/test_auto_publish.py
