+++ pulpcore/tests/functional/api/using_plugin/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for core functionality that require plugin involvement to exercise."""
