+++ pulpcore/tests/functional/api/using_plugin/test_content_delivery.py
