+++ pulpcore/tests/functional/api/using_plugin/test_purge.py
